requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
coverage==4.5.1
Django==2.1.2
django-bootstrap3==11.0.0
django-debug-toolbar==1.11
django-environ==0.4.5
django-filter==2.1.0
djangorestframework==3.9.1
flake8==3.6.0
flake8-django==0.0.3
Markdown==3.0.1
mccabe==0.6.1
more-itertools==5.0.0
pathlib==1.0.1
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytest==4.1.0
pytz==2018.7
six==1.12.0
sqlparse==0.2.4
